Testing the DELETE operation now also creates an account to ensure that a delete operation will always happen

+++ backend/model/crud_test.go
@@ -39,7 +39,33 @@ func TestDeleteUserAccountByEmail(t *testing.T) {
 		return
 	}
 
-	_, err := DeleteUserAccountByEmail(emailForTesting)
+	// Main logic for this test:
+	// 1 - Check if the test account exists
+	// 2 - If does not exist, create it
+	// 3 - If it already exists, delete it
+
+	doesUserExists := DoesUserAlreadyExists(emailForTesting)
+	if doesUserExists == true {
+		_, err := DeleteUserAccountByEmail(emailForTesting)
+		if err != nil {
+			t.Errorf("Could not delete the account, error: %s", err)
+		}
+		return
+	}
+
+	// Since the user doesn't exist at this point,
+	// it neeeds to be created
+	// The focus of this test is the DELETING operation, so only the email is needed
+	var user model.UserAccount
+	user.Email = emailForTesting
+
+	_, err := CreateUserAccount(&user)
+	if err != nil {
+		t.Errorf("Could not create an account so it could be deleted later, error: %s", err)
+		return
+	}
+
+	_, err = DeleteUserAccountByEmail(emailForTesting)
 	if err != nil {
 		t.Errorf("Could not delete the account, error: %s", err)
 	}
